Edit profile mod

+++ app/src/main/java/com/app/ecarepro/data/network/model/post_leave_request/LeaveRequestData.kt
@@ -1,7 +1,7 @@
 package com.app.ecarepro.data.network.model.post_leave_request
 
 data class LeaveRequestData(
-    val duration: Int,
+    val duration: Double,
     val fileAttachment: FileAttachment,
     val fromDate: String,
     val halfdayDTL: List<HalfdayDTL>,
